Moving namespace also.
